+++ bluetti_mqtt/core/devices/struct.py
@@ -13,10 +13,9 @@ def swap_bytes(data: bytes):
 
 
 class DeviceField:
-    def __init__(self, name: str, page: int, offset: int, size: int):
+    def __init__(self, name: str, address: int, size: int):
         self.name = name
-        self.page = page
-        self.offset = offset
+        self.address = address
         self.size = size
 
     def parse(self, data: bytes) -> Any:
@@ -27,9 +26,9 @@ class DeviceField:
 
 
 class UintField(DeviceField):
-    def __init__(self, name: str, page: int, offset: int, range: Optional[Tuple[int, int]]):
+    def __init__(self, name: str, address: int, range: Optional[Tuple[int, int]]):
         self.range = range
-        super().__init__(name, page, offset, 1)
+        super().__init__(name, address, 1)
 
     def parse(self, data: bytes) -> int:
         return struct.unpack('!H', data)[0]
@@ -42,17 +41,17 @@ class UintField(DeviceField):
 
 
 class BoolField(DeviceField):
-    def __init__(self, name: str, page: int, offset: int):
-        super().__init__(name, page, offset, 1)
+    def __init__(self, name: str, address: int):
+        super().__init__(name, address, 1)
 
     def parse(self, data: bytes) -> bool:
         return struct.unpack('!H', data)[0] == 1
 
 
 class EnumField(DeviceField):
-    def __init__(self, name: str, page: int, offset: int, enum: Type[Enum]):
+    def __init__(self, name: str, address: int, enum: Type[Enum]):
         self.enum = enum
-        super().__init__(name, page, offset, 1)
+        super().__init__(name, address, 1)
 
     def parse(self, data: bytes) -> Any:
         val = struct.unpack('!H', data)[0]
@@ -60,10 +59,10 @@ class EnumField(DeviceField):
 
 
 class DecimalField(DeviceField):
-    def __init__(self, name: str, page: int, offset: int, scale: int, range: Optional[Tuple[int, int]]):
+    def __init__(self, name: str, address: int, scale: int, range: Optional[Tuple[int, int]]):
         self.scale = scale
         self.range = range
-        super().__init__(name, page, offset, 1)
+        super().__init__(name, address, 1)
 
     def parse(self, data: bytes) -> Decimal:
         val = Decimal(struct.unpack('!H', data)[0])
@@ -77,9 +76,9 @@ class DecimalField(DeviceField):
 
 
 class DecimalArrayField(DeviceField):
-    def __init__(self, name: str, page: int, offset: int, size: int, scale: int):
+    def __init__(self, name: str, address: int, size: int, scale: int):
         self.scale = scale
-        super().__init__(name, page, offset, size)
+        super().__init__(name, address, size)
 
     def parse(self, data: bytes) -> Decimal:
         values = list(struct.unpack(f'!{self.size}H', data))
@@ -99,8 +98,8 @@ class SwapStringField(DeviceField):
 
 
 class VersionField(DeviceField):
-    def __init__(self, name: str, page: int, offset: int):
-        super().__init__(name, page, offset, 2)
+    def __init__(self, name: str, address: int):
+        super().__init__(name, address, 2)
 
     def parse(self, data: bytes) -> int:
         values = struct.unpack('!2H', data)
@@ -108,8 +107,8 @@ class VersionField(DeviceField):
 
 
 class SerialNumberField(DeviceField):
-    def __init__(self, name: str, page: int, offset: int):
-        super().__init__(name, page, offset, 4)
+    def __init__(self, name: str, address: int):
+        super().__init__(name, address, 4)
 
     def parse(self, data: bytes) -> int:
         values = struct.unpack('!4H', data)
@@ -122,47 +121,47 @@ class DeviceStruct:
     def __init__(self):
         self.fields = []
 
-    def add_uint_field(self, name: str, page: int, offset: int, range: Tuple[int, int] = None):
-        self.fields.append(UintField(name, page, offset, range))
+    def add_uint_field(self, name: str, address: int, range: Tuple[int, int] = None):
+        self.fields.append(UintField(name, address, range))
 
-    def add_bool_field(self, name: str, page: int, offset: int):
-        self.fields.append(BoolField(name, page, offset))
+    def add_bool_field(self, name: str, address: int):
+        self.fields.append(BoolField(name, address))
 
-    def add_enum_field(self, name: str, page: int, offset: int, enum: Type[Enum]):
-        self.fields.append(EnumField(name, page, offset, enum))
+    def add_enum_field(self, name: str, address: int, enum: Type[Enum]):
+        self.fields.append(EnumField(name, address, enum))
 
-    def add_decimal_field(self, name: str, page: int, offset: int, scale: int, range: Tuple[int, int] = None):
-        self.fields.append(DecimalField(name, page, offset, scale, range))
+    def add_decimal_field(self, name: str, address: int, scale: int, range: Tuple[int, int] = None):
+        self.fields.append(DecimalField(name, address, scale, range))
 
-    def add_decimal_array_field(self, name: str, page: int, offset: int, size: int, scale: int):
-        self.fields.append(DecimalArrayField(name, page, offset, size, scale))
+    def add_decimal_array_field(self, name: str, address: int, size: int, scale: int):
+        self.fields.append(DecimalArrayField(name, address, size, scale))
 
-    def add_string_field(self, name: str, page: int, offset: int, size: int):
-        self.fields.append(StringField(name, page, offset, size))
+    def add_string_field(self, name: str, address: int, size: int):
+        self.fields.append(StringField(name, address, size))
 
-    def add_swap_string_field(self, name: str, page: int, offset: int, size: int):
-        self.fields.append(SwapStringField(name, page, offset, size))
+    def add_swap_string_field(self, name: str, address: int, size: int):
+        self.fields.append(SwapStringField(name, address, size))
 
-    def add_version_field(self, name: str, page: int, offset: int):
-        self.fields.append(VersionField(name, page, offset))
+    def add_version_field(self, name: str, address: int):
+        self.fields.append(VersionField(name, address))
 
-    def add_sn_field(self, name: str, page: int, offset: int):
-        self.fields.append(SerialNumberField(name, page, offset))
+    def add_sn_field(self, name: str, address: int):
+        self.fields.append(SerialNumberField(name, address))
 
-    def parse(self, page: int, offset: int, data: bytes) -> dict:
+    def parse(self, starting_address: int, data: bytes) -> dict:
         # Offsets and size are counted in 2 byte chunks, so for the range we
         # need to divide the byte size by 2
         data_size = int(len(data) / 2)
 
         # Filter out fields not in range
-        r = range(offset, offset + data_size)
+        r = range(starting_address, starting_address + data_size)
         fields = [f for f in self.fields
-                  if f.page == page and f.offset in r and f.offset + f.size - 1 in r]
+                  if f.address in r and f.address + f.size - 1 in r]
 
         # Parse fields
         parsed = {}
         for f in fields:
-            data_start = 2 * (f.offset - offset)
+            data_start = 2 * (f.address - starting_address)
             field_data = data[data_start:data_start + 2 * f.size]
             val = f.parse(field_data)